CABIssueLinkage/requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
decorator==3.4.0
ipython==2.3.1
networkx==1.9.1
numpy==1.9.1
python-dateutil==2.4.0
pyzmq==14.4.1
requests==2.5.1
scipy==0.15.0
six==1.9.0
tornado==4.0.2
wsgiref==0.1.2
